chore: Refactor hook execution.
