+++ packages/shared/sdk-client/tsconfig.json
@@ -16,6 +16,6 @@
     "resolveJsonModule": true,
     "types": ["jest", "node"]
   },
-  "include": ["src"],
+  "include": ["src", "../sdk-server/src/hooks"],
   "exclude": ["**/*.test.ts", "dist", "node_modules", "__tests__"]
 }
